refactor: Temporarily disable HomePreview in MainActivity
This commit comments out the `HomePreview` composable in `MainActivity.kt`. This preview was used to show a preview of the `HomeView` in the design view within Android Studio.

This change is temporary and the preview will be re-enabled in a subsequent commit once necessary dependencies and configurations are in place.